backend/requirements.txt
FastAPI
SQLAlchemy
pandas
geopy
uvicorn
aiofiles
python-dotenv
requests
psycopg2
